requirements.txt
opencv-python
visdom
dominate
numpy
scipy
pillow
unidecode
requests
tqdm
